add appearance new prop
