+++ lib/schemas/browse/modules/components/asyncWrapper.js
@@ -23,5 +23,5 @@ module.exports = makeComponent({
 		field: { $ref: 'schemaDefinitions#/definitions/browseField' },
 		targetField: { type: 'string' }
 	},
-	requiredProperties: ['dataMapping', 'source', 'field']
+	requiredProperties: ['source', 'field']
 });
