+++ lib/schemas/browse/modules/field.js
@@ -5,42 +5,8 @@ const components = require('./components');
 const mapper = require('../../common/mapper');
 const makeConditions = require('../../common/conditions');
 const interactions = require('./interactions');
+const appearance = require('./field-appearance');
 const filter = require('./filter');
-
-const makeAppearance = () => {
-	const appearanceProps = {
-		type: 'object',
-		properties: {
-			fontColor: { type: 'string' },
-			fontSize: {
-				enum: [
-					'base',
-					'baseSmall',
-					'small',
-					'xsmall',
-					'large',
-					'xlarge',
-					'xxlarge'
-				]
-			},
-			align: { enum: ['left', 'center', 'right'] },
-			verticalAlign: { enum: ['top', 'center', 'bottom'] }
-		},
-		additionalProperties: false,
-		minProperties: 1
-	};
-
-	return {
-		type: 'object',
-		properties: {
-			default: appearanceProps,
-			desktop: appearanceProps,
-			mobile: appearanceProps
-		},
-		additionalProperties: false,
-		minProperties: 1
-	};
-};
 
 module.exports = {
 	type: 'object',
@@ -50,7 +16,7 @@ module.exports = {
 		label: { type: 'string' },
 		noLabel: { type: 'boolean' },
 		deviceDisplay: { enum: ['desktop', 'mobile'] },
-		appearance: makeAppearance(),
+		appearance,
 		attributes: {
 			type: 'object',
 			properties: {
